requirements.txt
# Core LangChain
langchain
langchain-core
langchain-community
langchain-huggingface
langchain-google-genai
langchain-qdrant

# Vector DB
qdrant-client

# Embeddings
torch
sentence-transformers

# PDF loading
pymupdf

# BM25 reranking
rank-bm25

# API & UI
fastapi
uvicorn[standard]
streamlit
python-multipart

# Utilities
python-dotenv